pyproject.toml
[project]
name = "web-compilaton-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.9"
dependencies = [
    "django>=6.0.2",
    "celery>=5.6.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "flower>=2.0.1",
    "requests>=2.32.5",
    "redis>=7.2.1",
]

[tool.pyright]
include=["backend"]
pythonVersion="3.12.9"
venvPath ="."
venv=".venv"
